chore: moved Error logging class outside of script renamed errors_to_fix to error_log_store
