+++ server/preprocess/add_descriptions_bird_dataset.py
@@ -7,6 +7,7 @@ from typing import Union
 import pandas as pd
 from tqdm import tqdm
 
+from preprocess.AddDescriptionErrorLogs import AddDescriptionErrorLogs
 from services.client_factory import Client, ClientFactory
 from utilities.config import PATH_CONFIG
 from utilities.constants.database_enums import DatasetType
@@ -24,13 +25,13 @@ from utilities.constants.preprocess.add_descriptions_bird_dataset.response_messa
     ERROR_COLUMN_MEANING_FILE_NOT_FOUND,
     ERROR_ENSURING_DESCRIPTION_FILES_EXIST,
     ERROR_FAILED_TO_READ_CSV,
+    ERROR_FILE_NOT_FOUND,
     ERROR_GENERATING_COLUMN_DESCRIPTIONS,
     ERROR_GENERATING_TABLE_DESCRIPTIONS,
     ERROR_INITIALIZING_DESCRIPTION_FILES,
+    ERROR_SQLITE_EXECUTION_ERROR,
     ERROR_TABLE_DOES_NOT_EXIST,
     ERROR_UPDATING_DESCRIPTION_FILES,
-    ERROR_SQLITE_EXECUTION_ERROR,
-    ERROR_FILE_NOT_FOUND,
     INFO_COLUMN_ALREADY_HAS_DESCRIPTIONS,
     INFO_TABLE_ALREADY_HAS_DESCRIPTIONS,
     WARNING_ENCODING_FAILED,
@@ -45,8 +46,8 @@ from utilities.prompts.prompt_templates import (
 from utilities.utility_functions import (
     format_schema,
     get_table_columns,
-    get_table_names,
     get_table_ddl,
+    get_table_names,
 )
 
 logger = setup_logger(__name__)
@@ -62,27 +63,6 @@ TABLE_DESCRIPTION_PLACEHOLDER = "No Description Available"
 
 PRAGMA_COLUMN_NAME_INDEX = 1
 PRAGMA_COLUMN_TYPE_INDEX = 2
-
-
-class ErrorsToFix:
-    """
-    A singleton class that maintains a shared list of errors to fix.
-
-    This class ensures that only one instance exists throughout the application.
-    It uses the singleton pattern by overriding the __new__ method.
-
-    Attributes:
-        errors (list): A list to store error messages or error objects.
-    """
-
-    _instance = None
-
-    def __new__(cls):
-        if cls._instance is None:
-            cls._instance = super().__new__(cls)
-            cls._instance.errors = []  # Initialize the list
-        return cls._instance
-
 
 def read_csv(
     file_path: str, encodings: list = ["utf-8-sig", "ISO-8859-1"]
@@ -227,7 +207,7 @@ def get_improved_column_description(
         str: An improved description of the column generated by the language model.
     """
 
-    errors_to_fix = ErrorsToFix()
+    error_log_store = AddDescriptionErrorLogs()
 
     improved_description = ""
     try:
@@ -256,7 +236,7 @@ def get_improved_column_description(
 
         improved_description = client.execute_prompt(prompt)
     except Exception as e:
-        errors_to_fix.errors.append(
+        error_log_store.errors.append(
             {
                 "database": database_name,
                 "error": ERROR_GENERATING_COLUMN_DESCRIPTIONS.format(
@@ -406,7 +386,7 @@ def improve_column_descriptions_for_table(
         pd.DataFrame: The input DataFrame updated with improved column descriptions.
     """
 
-    errors_to_fix = ErrorsToFix()
+    error_log_store = AddDescriptionErrorLogs()
 
     # Get column names from the SQLite database for validation
     column_names = get_table_columns(connection, table_name)
@@ -452,7 +432,7 @@ def improve_column_descriptions_for_table(
                         column_name=row[ORIG_COLUMN_NAME_COL], table_name=table_name
                     ),
                 }
-                errors_to_fix.errors.append(error)
+                error_log_store.errors.append(error)
                 logger.error(error["error"])
         else:
             logger.info(
@@ -491,7 +471,7 @@ def improve_column_descriptions(
         RuntimeError: If an unexpected error occurs during processing.
     """
 
-    errors_to_fix = ErrorsToFix()
+    error_log_store = AddDescriptionErrorLogs()
 
     try:
 
@@ -518,7 +498,7 @@ def improve_column_descriptions(
                     database_name=database_name,
                 )
                 if table_not_in_db is not None:
-                    errors_to_fix.errors.append(table_not_in_db)
+                    error_log_store.errors.append(table_not_in_db)
                     continue
 
                 table_df = read_csv(os.path.join(
@@ -537,7 +517,7 @@ def improve_column_descriptions(
                     database_description_path, table_column_description_csv), index=False)
 
     except Exception as e:
-        errors_to_fix.errors.append(
+        error_log_store.errors.append(
             {"database": database_name, "error": f"{str(e)}"})
         raise RuntimeError(str(e))
 
@@ -568,7 +548,7 @@ def create_database_tables_csv(
     Raises:
         RuntimeError: If any unexpected error occurs during the generation or file writing process.
     """
-    errors_to_fix = ErrorsToFix()
+    error_log_store = AddDescriptionErrorLogs()
 
     try:
         table_description_csv_path = PATH_CONFIG.table_description_file(
@@ -615,7 +595,7 @@ def create_database_tables_csv(
                 table_description = client.execute_prompt(
                     table_description_prompt)
             except Exception as e:
-                errors_to_fix.errors.append(
+                error_log_store.errors.append(
                     {
                         "database": database_name,
                         "error": ERROR_GENERATING_TABLE_DESCRIPTIONS.format(
@@ -640,7 +620,7 @@ def create_database_tables_csv(
 
     except Exception as e:
 
-        errors_to_fix.errors.append(
+        error_log_store.errors.append(
             {"database": database_name, "error": f"{str(e)}"})
         raise RuntimeError(str(e))
 
@@ -870,7 +850,7 @@ def add_database_descriptions(
         Exception: If any error occurs during the processing of a database, it is logged and the exception is raised.
     """
 
-    errors_to_fix = ErrorsToFix()
+    error_log_store = AddDescriptionErrorLogs()
     dataset_dir = PATH_CONFIG.dataset_dir(dataset_type=dataset_type)
 
     client = ClientFactory.get_client(llm_type, model, temperature, max_tokens)
@@ -900,7 +880,7 @@ def add_database_descriptions(
                     client=client,
                     connection=connection,
                 )
-                if len(errors_to_fix.errors) == 0:
+                if len(error_log_store.errors) == 0:
                     improve_column_descriptions(
                         database_name=database,
                         dataset_type=dataset_type,
@@ -912,8 +892,8 @@ def add_database_descriptions(
             logger.error(
                 f"Error processing database {database}: {e}", exc_info=True)
 
-    if errors_to_fix.errors:
-        for error in errors_to_fix.errors:
+    if error_log_store.errors:
+        for error in error_log_store.errors:
             logger.error(
                 f"- Database: {error['database']}, Error: {error['error']}")
     else:
